==> 2_base_anac.ipynb <==
%pip install unidecode
from unidecode import unidecode
import pandas as pd

#Limpeza da base de AERODROMOS PUBLICOS (tirar as colunas indesejadas)

pd.set_option('display.max_columns', None)

caminho_arquivo = 'data/csv/pub.csv'  
df = pd.read_csv(caminho_arquivo, encoding="utf-8", sep=";", skiprows=1)

colunas_para_excluir = ['OACI', 'REFERÊNCIA(S)', 'RAMPA(S) DE APROXIMAÇÃO', 'RESISTÊNCIA.3', 'SUPERFÍCIE.3','AIP BRASIL', 'p1','p2','p3','p4','FORMATO DA ÁREA DE POUSO', 'PORTARIAS','Unnamed: 36','Unnamed: 37','Unnamed: 38', 'r1', 'r2', 'r1.1', 'REFERÊNCIA', 'SITUAÇÃO', 'RESTRIÇÃO(ÕES)', 'OBSERVAÇÕES']
df = df.drop(columns=colunas_para_excluir)

indices_para_excluir = [30]  # Substitua pelos índices das colunas que deseja remover
df = df.drop(df.columns[indices_para_excluir], axis=1)

indices_para_excluir = [27]  # Substitua pelos índices das colunas que deseja remover
df = df.drop(df.columns[indices_para_excluir], axis=1)

indices_para_excluir = [26]  # Substitua pelos índices das colunas que deseja remover
df = df.drop(df.columns[indices_para_excluir], axis=1)

indices_para_excluir = [25]  # Substitua pelos índices das colunas que deseja remover
df = df.drop(df.columns[indices_para_excluir], axis=1)

indices_para_excluir = [24]  # Substitua pelos índices das colunas que deseja remover
df = df.drop(df.columns[indices_para_excluir], axis=1)
indices_para_excluir = [26]  # Substitua pelos índices das colunas que deseja remover
df = df.drop(df.columns[indices_para_excluir], axis=1)

for i, col in enumerate(df.columns):
    print(f"{i}: {col}")


 #Arrumando os nomes
def rename_duplicate_columns(df):
    new_columns = []
    
    # Lista das colunas que precisam do sufixo "_Pista1"
    colunas_pista1 = {'DESIGNAÇÃO', 'COMPRIMENTO', 'LARGURA', 'RESISTÊNCIA', 'SUPERFÍCIE'}
    
    for col in df.columns:
        if col in colunas_pista1:
            new_columns.append(col + "_Pista1")  # Adiciona _Pista1 às colunas específicas
        elif ".1" in col:
            new_columns.append(col.replace(".1", "_Pista2"))  # Substitui .1 por _Pista2
        elif ".2" in col:
            new_columns.append(col.replace(".2", "_Pista3"))  # Substitui .2 por _Pista3
        else:
            new_columns.append(col)  # Mantém as colunas que não precisam de alteração

    df.columns = new_columns
    return df

df = rename_duplicate_columns(df)

df = df.rename(columns={"MUNICÍPIO ATENDIDO": "municipio"})
df = df.rename(columns={"CÓDIGO OACI": "codigo_oaci"})
df2 = df


df2["classificacao"] = "publico"

print(df2.columns)
for i, col in enumerate(df2.columns):
    print(f"{i}: {col}")


#lendo a segunda base AERODROMOS PRIVADOS 

priv = pd.read_csv('data/csv/priv.csv', sep=';', encoding='latin1', skiprows=1)  
pd.set_option('display.max_columns', None)

indices_para_excluir_priv = [20,21,22,23]  # Substitua conforme necessário
colunas_para_excluir_por_indice = [priv.columns[i] for i in indices_para_excluir_priv if i < len(priv.columns)]
priv = priv.drop(columns=colunas_para_excluir_por_indice)

def rename_columns_priv(df):
    df = df.rename(columns=lambda x: x.strip().replace(" ", "_"))  # Remove espaços e substitui por _
    df = df.rename(columns=lambda x: x.replace("1", "_Pista1").replace("2", "_Pista2"))  # Substitui 1 e 2
    return df

# Aplicar a função na base priv
priv = rename_columns_priv(priv)

priv["classificacao"] = "privado"
priv.columns = priv.columns.str.replace('__', '_', regex=True)
priv2 = priv

# Função para padronizar a coluna operação
priv2["operacao"] = priv2.apply(lambda x: f"{x['Operação_Diurna']} Diurno e {x['Operação_Noturna']} Noturno", axis=1)
priv2 = priv2.drop(columns=["Operação_Diurna", "Operação_Noturna"])

for i, col in enumerate(priv2.columns):
    print(f"{i}: {col}")

   

#padronizar as bases 

# Padronizar os nomes das colunas (sem acentos e tudo em minúsculas)
priv2.columns = [unidecode(col).lower() for col in priv2.columns]
df2.columns = [unidecode(col).lower() for col in df2.columns]

# Exibir os nomes das colunas para conferência
print("Colunas de df2:", df2.columns.to_list())
print("Colunas de priv2:", priv2.columns.to_list())

# Juntar as bases
aero_ANAC = pd.concat([df2, priv2], ignore_index=True).fillna("N/A")

# Conferir os totais de registros
total_df2 = len(df2)
total_priv2 = len(priv2)
total_aero_ANAC = len(aero_ANAC)

print(f"Total de registros em df2: {total_df2}")
print(f"Total de registros em priv2: {total_priv2}")
print(f"Total de registros em aero_ANAC: {total_aero_ANAC}")

# Exibir a contagem de valores na coluna 'classificacao'
if "classificacao" in aero_ANAC.columns:
    print(aero_ANAC["classificacao"].value_counts())
else:
    print("A coluna 'classificacao' não existe em aero_ANAC.")

# Lista de superfícies que devem ter a operação como "VFR"
superficies_vfr = ["Grama", "Terra", "Cascalho", "Piçarra", "Barro", "Areia", "Saibro"]

# Atualizar os valores da coluna "Operacao" para "VFR" quando qualquer pista for uma das superfícies listadas
aero_ANAC.loc[
    aero_ANAC[["superficie_pista1", "superficie_pista2", "superficie_pista3"]].isin(superficies_vfr).any(axis=1),
    "operacao"
] = "VFR"

print(df.columns.to_list())

print(priv.columns.to_list())

print(aero_ANAC.columns.to_list())

# Salvar a base aero_ANAC como CSV
aero_ANAC.to_csv("output/aero_ANAC.csv", index=False, encoding="utf-8")
print("Arquivo 'aero_ANAC.csv' salvo com sucesso!")
